# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "2.0.0",
    "title": "WPPConnect API Rest",
    "description": "Welcome to the wppconnect-server API documentation. This API provides a set of endpoints to interact with the wppconnect-server application, allowing you to build integrations and automate interactions with WhatsApp."
  },
...
  "paths": {
...
    "/api/{secretkey}/show-all-sessions": {
      "get": {
        "tags": ["Auth"],
        "description": "",
        "operationId": "showAllSessions",
        "parameters": [
          {
            "name": "secretkey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "THISISMYSECURETOKEN"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/{secretkey}/start-all": {
      "post": {
        "tags": ["Auth"],
        "description": "",
        "operationId": "startAllSessions",
        "parameters": [
          {
            "name": "secretkey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "THISISMYSECURECODE"
            }
          },
          {
            "name": "session",
            "schema": {
              "type": "string",
              "example": "NERDWHATS_AMERICA"
            },
            "in": "query"
          },
          {
            "name": "authorization",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/{session}/check-connection-session": {
      "get": {
        "tags": ["Auth"],
        "description": "",
        "operationId": "CheckConnectionState",
        "parameters": [
          {
            "name": "session",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "NERDWHATS_AMERICA"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/{session}/get-media-by-message/{messageId}": {
      "get": {
        "tags": ["Messages"],
        "description": "",
        "operationId": "getMediaByMessage",
        "parameters": [
          {
            "name": "session",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "NERDWHATS_AMERICA"
            }
          },
          {
            "name": "messageId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/{session}/qrcode-session": {
      "get": {
        "tags": ["WA-JS"],
        "description": "",
        "operationId": "getQRCode",
        "parameters": [
          {
            "name": "session",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "NERDWHATS_AMERICA"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": []
      }
    },
    "/api/{session}/start-session": {
      "post": {
        "tags": ["WA-JS"],
        "description": "",
        "operationId": "startSession",
        "parameters": [
          {
            "name": "session",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "NERDWHATS_AMERICA"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": []
      }
    },
    "/api/{session}/logout-session": {
      "post": {
        "tags": ["WA-JS"],
        "description": "",
        "operationId": "logoutSession",
        "parameters": [
          {
            "name": "session",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "NERDWHATS_AMERICA"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": []
      }
    },
...
    "/api/{session}/close-session": {
      "post": {
        "tags": ["WA-JS"],
        "description": "",
        "operationId": "closeSession",
        "parameters": [
          {
            "name": "session",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "NERDWHATS_AMERICA"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": []
      }
    },
    "/api/{session}/subscribe-presence": {
      "post": {
        "tags": ["Misc"],
        "description": "",
        "operationId": "subscribePresence",
        "parameters": [
          {
            "name": "session",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "NERDWHATS_AMERICA"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "phone": {
                    "type": "string"
                  },
                  "isGroup": {
                    "type": "boolean"
                  },
                  "all": {
                    "type": "boolean"
                  }
                }
              },
              "example": {
                "phone": "5521999999999",
                "isGroup": false,
                "all": false
              }
            }
          }
        }
      }
    },
    "/api/{session}/set-online-presence": {
      "post": {
        "tags": ["Misc"],
        "description": "",
        "operationId": "setOnlinePresence",
        "parameters": [
          {
            "name": "session",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "NERDWHATS_AMERICA"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "isOnline": {
                    "type": "boolean"
                  }
                }
              },
              "example": {
                "isOnline": false
              }
            }
          }
        }
      }
    },
    "/api/{session}/download-media": {
      "post": {
        "tags": ["Messages"],
        "description": "",
        "operationId": "downloadMediabyMessage",
        "parameters": [
          {
            "name": "session",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "example": "NERDWHATS_AMERICA"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "messageId": {
                    "type": "string"
                  }
                }
              },
              "example": {
                "messageId": "<messageId>"
              }
            }
          }
        }
      }
    },
...
    "/api/{session}/all-groups": {
      "get": {
        "tags": ["WA-JS"],
        "description": "",
        "operationId": "getAllGroups",
        "parameters": [
          {
            "name": "session",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": []
      }
    },
...
    "/api/{session}/all-contacts": {
      "get": {
        "tags": ["WA-JS"],
        "description": "",
        "operationId": "getAllContacts",
        "parameters": [
          {
            "name": "session",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}